Fix tx_st_empty
