+++ ip/fejkon_pcie_data/Makefile
@@ -5,7 +5,7 @@
 SRCS := ../pcie_msi_intr/pcie_msi_intr.sv fejkon_pcie_data.sv test.sv
 
 lint: $(SRCS)
-	pylint3 --rcfile=pylintrc test.py --exit-zero
+	pylint3 --rcfile=pylintrc test.py test_ep.py --exit-zero
 	verilator --lint-only fejkon_pcie_data.sv
 	iverilog -g2012 -t null $^
